fix:created a widget for the batch detail container
