+++ lib/features/farms report/presentation/screens/chicken_reduction_page.dart
@@ -1,7 +1,8 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:ikuku/features/batches/model/chicken_batch_model.dart';
-import 'package:ikuku/features/farms%20report/presentation/widgets/reducion_reasons_checkboxes.dart';
+import 'package:ikuku/features/farms%20report/presentation/widgets/batch_detail_container.dart';
+import 'package:ikuku/features/farms%20report/presentation/widgets/reduction_reasons_checkboxes.dart';
 import 'package:ikuku/shared/widgets/feature_button.dart';
 import 'package:ikuku/theme/app_theme.dart';
 
@@ -51,42 +52,7 @@ class _ChickenReductionPageState extends State<ChickenReductionPage> {
               
               // Batch Detail Container Card
               if (widget.batch != null)
-                Container(
-                  width: double.infinity,
-                  padding: const EdgeInsets.all(16),
-                  margin: const EdgeInsets.only(bottom: 24),
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(16),
-                    border: Border.all(color: Colors.grey.shade600, width: 1.8),
-                  ),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-                        'selected_batch'.tr(),
-                        style: const TextStyle(
-                          fontSize: 14,
-                          color: Colors.grey,
-                          fontWeight: FontWeight.w500,
-                        ),
-                      ),
-                      const SizedBox(height: 8),
-                      Text(
-                        widget.batch!.name,
-                        style: const TextStyle(
-                          fontSize: 18,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      const SizedBox(height: 6),
-                      Text(
-                        '${widget.batch!.typeOfBird.toUpperCase()} • ${widget.batch!.initialCount} ${'birds'.tr()} • ${widget.batch!.age} ${widget.batch!.ageUnit} ${'old'.tr()}',
-                        style: const TextStyle(fontSize: 14, color: Colors.grey),
-                      ),
-                    ],
-                  ),
-                ),
+                BatchDetailContainer(batch: widget.batch!),
       
               const SizedBox(height: 8),
               Column(
